spark_etl/requirements.txt
pyspark
requests
psycopg2-binary
pytest